requirements.txt
fastapi
uvicorn
requests
twilio
python-dotenv
groq
